+++ src/main/kotlin/io/openfuture/chain/domain/crypto/MasterKeyRequest.kt
@@ -1,5 +1,0 @@
-package io.openfuture.chain.domain.crypto
-
-data class MasterKeyRequest(
-        val seedPhrase: String
-)