+++ src/main/kotlin/io/openfuture/chain/domain/crypto/RootAccountDto.kt
@@ -1,0 +1,9 @@
+package io.openfuture.chain.domain.crypto
+
+import io.openfuture.chain.domain.crypto.key.KeyDto
+
+data class RootAccountDto(
+    val seedPhrase: String,
+    val masterKeys: KeyDto,
+    val defaultAccount: AccountDto
+)