+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Address validation
 - RPC requests: validateAddress
 
+### Changed
+- Default account info is added to the doGenerateMaster end-point response
+- doGenerateMaster end-point renamed to doRestore
+
 ## [0.2.0] - 2018-06-29
 ### Added
 - Seed phrase generator